Move project index syntax error formatter into shared reporting
