+++ src/shared/reporting/index.js
@@ -1,1 +1,2 @@
 export * from "./metrics-tracker.js";
+export * from "./syntax-error-formatter.js";
