+++ src/shared/reporting/syntax-error-formatter.js
@@ -1,8 +1,15 @@
 import path from "node:path";
 
-import { getNonEmptyString } from "../../../shared/string-utils.js";
-import { resolveContainedRelativePath } from "../../../shared/path-utils.js";
+import { getNonEmptyString } from "../string-utils.js";
+import { resolveContainedRelativePath } from "../path-utils.js";
 
+/**
+ * Format parser-originated syntax errors into the structured messages surfaced
+ * by the CLI and project index. Centralising the formatter in the shared
+ * reporting layer keeps downstream packages from re-implementing the display
+ * logic or hard-coding project-index paths when converting absolute file
+ * locations into friendly relative references.
+ */
 export function formatProjectIndexSyntaxError(error, sourceText, context) {
     const { filePath, projectRoot } = context ?? {};
     const lineNumber = getFiniteNumber(error.line);
